Addes scripts for re-mounting system default mount points as rw or ro
It is also possible to specify expicit mount points, bypassing the
local system default list of mount points to be affected by the scripts.

The main purpose for the new scripts is to re-mount /usr, /opt etc.
before and after installation of new Portage packages, so that
those mount points are normally mounted as read-only.