Added script for determining commenting conventions from file extension
The new script checks the specified file extensions for a list
of known extension and reports information such as "open comment" and
"close comment" tokens.

A file name can also be specified instead of a file extension; in this
case the file extension of the filename will be used, or the file
contents will be analyzed (if it exists).