Repository-relative paths will start with ./ now in the ignore list
Before, paths were prefixed with "/" instead of "./".

While this works with the current versions of Bazaar-NG, it still seems
a bit dangerous to me, and so I used the less dangerous variant.

BTW, the old paths in the ignore list will be converted silently
into the new format if any new entries are found.