Made remount-readonly warnings the default
Until now, they have just been silently ignored unless a --warning
option was provided.

But as remounting /usr as read-only fails so often due to updating or
prelinking a library while it was in use, the default now is to always
display warnings and a --quiet option was added instead to revert to
the old default.

Also added some hints how to use lsof in order to find out which
program is blocking remounting as read-only.