Actually implemented --quick-check option for wait-for-network-idle
The help text already has referenced it for the longest time anyway.

Also, the possible return code 1 has been restricted to be allowed
if the --quick-check option has been specified only.

This allows the script to be used with && without caring whether
or not then network has been idle after the initial guard time.