Patterns are always matched against the original file now
Duplicate elimination prefers checksum files over original files, but
the user-specified patterns are not targetted at checksum files.

Therefore, they would not match.

In order to fix this, the pattern match now always removes the checksum
file extension. This allows the original files to match the same
patterns as the checksum files.