Simplified option handling for niceness settings
The old fork-approach required a more complicated option handling. This
is no longer necessary now.

Also changed the default I/O niceness to "best effort" class 2 with
priority 6.

Priority 7 would be the lowest priority in this class, but I wanted a
chance to assign an even lower priority to other I/O jobs.

Class 3 "idle" was probably a bad idea - this should really be reserved
to jobs where the completion duration is not important. This need not
necessarily be true when running the checksummer, however, and so the
default priority was changed.