Added script for "invisibly" displaying text
It is a filter using the same foreground and background color for
output.

This will not be visible on an ANSI-terminal, but the mouse can still
select the "invisible" text.

The script is intended for displaying passwords which are then copied
with the mouse. This way nobody looking over the shoulder can get a
look on the password.