Include branch nicknames in generated merge directives
The nicknames are also included within the files storing the last
known commits. They are not used by the script, but rather for the
user's benefit: The user can determine the branch to which the file
belongs.

The file names are just hashes, so they do not tell much by themselves.