Enhanced IPv4 synthesizer to optionally extract FQDNs from /etc/hosts
This is only done if no arguments have been specified.

Also, only such FQDNs are taken from /etc/hosts which currently refer to
IPv4 addresses in the ranges reserved for local nets.

In addition, as a quick shortcut, also such FQDNs are taken which refer
to an IPv4 address starting with "0.", typically. "0.0.0.0".

This allows to manually prepare entries in /etc/hosts for extraction by
the script by changing the IPv4 addresses to such a 0.*-based value,
even for IP addresses currently outside of the reserved private range (
changing the first component to "0" is easier than changing it to,
say "10").

Finally, the script has been renamed in order to communicate that it only
supports IPv4 addresses, not IPv6 addresses. (Everyone is talking about
migration to IPv6 these days. I really don't care - not within my LANs!)