Changed behaviour when uninstalling symlinks to directories
rmdir instead of rm will be used in this case, which will typically fail
and not remove the directory.

However, removing the symlink in such cases could be dangerous - just
think of a symlink /usr64 -> /usr and the package was installed in /usr.
One certainly would not like the symlink /usr to be removed in this
case!

Also, help text were added to configure4user and uninstall-package-list.

The -h option will display it.