Started converting the uninstallation script into a package manager
The help describes things not implemented yet.

The -u command is the only thing that works, and performs the same
actions as the old uninstallation script. (Despite of the new
functionality described in the help.)

The worst problem of the old functionality is that installation
overwrote the permissions of already existing directories with those of
the staging directory.