Added script for converting text into QEMU monitor "sendkey" commands
This allows to send the keystrokes representing a very long password to
a qemu guest by pasting the output of this command into the qemu monitor
console of the guest vm.

In order to be able to paste the text to the monitor, redirect it via
TCP or UNIX domain sockets to a local terminal where you can paste.

For instance when using the following option as part of the qemu
command line

-monitor "unix:`pwd`/monitor,server,nowait"

then a UNIX socket ./monitor represents the monitor console on the host
where qemu is tunning.

This can be tunneled from that host to your local machine by using a
tool like socat:

$ echo "ssh qemu.host.com socat /path/to/monitor/socket STDIO" > redir
$ socat EXEC:'sh ./redir' UNIX-LISTEN:./local_end &
$ rlwrap unixterm ./local_end

Paste the output of the new script to the unixterm console. This will
send the keystrokes to the guest.

When you are done, press ctrl-d to close the monitor connection and
remove the local helper script and helper socket:

$ rm redir local_end