Migrated expand-series.guards to work with dash
The actual problem was even bigger: It turned out that the dot command
in dash does not accept any arguments beyond the script name!

In bash and ash, the dot command can have any number of arguments which
are treated in exactly the same way as arguments to a user-defined
function invocation.

Therefore, the functions/lookup.sh source snippet hat to be modified as
well in order to process $@ of the invoking script rather than a
private version of it.