Initial version of TCP receive/send script branch
This branch provides scripts to send or receive arbitrary files directly
over TCP using a specific network interface.

Usage is simple: Start the receive*-script on the receiving side, then
use the send*-script of the sending side to actually commence the
transfer.

Normally the receive* script listens on the local machine for
connections made by the send* script from another machine. In this case,
no argument is given to the receive* script.

But it is also possible to specify a host name or IP for the receive*
script, in which case it assumes the sender is listening for connections
and fetches the files from there.

The files themselves are put in a TAR archive on the fly, but this is
transparent to the user.

Then the "socat" utility (which needs to be installed) is used to
transfer the archive over the network.

The scripts are therefore just wrappers around tar and socat. The
advantage is easier use, because the scripts allow to "hard-wire" the
interface name and port by creating symlinks.

The original scripts (currently) have the names

*  receive_files_as_tar_from_lo_port_9999

*  send_files_as_tar_to_lo_port_9999

where "lo" is the interface (as displayed by /sbin/ifconfig) and "9999"
is the port number. The scripts parse those values from the script name,
any by creating symlinks such as "send_files_as_tar_to_eth0_port_12345"
different interfaces or ports can be used.

The scripts actually bind to IP addresses rather than to interfaces, but
the IP address is looked up automatically from the interface name by the
scripts. This has the same effect as binding to the interface.

Binding to specific interfaces is important as a security measure,
because the tar archive is transferred unencrypted in the clear to
anyone which can connect to the port on the other side.

The scripts are intended to be used in a local LAN where no attackers
are present to "steal" the transferred files, and therefore the
transfers should be bound to interfaces which are only reachable from
within the LAN. (If you are behind a firewall, this is not important.)

The scripts are somewhat complementary to scp and sftp in that they do a
similar job, but without the encryption overhead (and security). They
are therefore better suited than scp/sftp if large amounts of data have
to be transferred and the transfer takes place inside a secure LAN
anyway.

Note that even though no authentication or encryption is provided by the
scripts, the internal checksums used by TCP as well as tar should
nevertheless provide sufficient protection against transmission errors.

Another advantage is that no ssh server is required on the remote host -
just the receive-script needs to be transferred there somehow (e. g. by
logging in by telnet and transferring as text via copy/paste).

This branch contains scripts originally developed as part of branch
trunk-twe61ijrbh6qpw776bv20l4su, but have been moved into a branch of
their own for the sake of easier merging into otherwise unrelated
account-specific branches.

This shared branch is intended to be merged into an account-specific
"bin" subdirectory with is in $PATH, such as "~/bin/local" or just
"~/bin".