Added --override-script-name option
This is useful if the script is exec'ed from another script rather than
being symlinked, because it still provides a way to "hard-wire"
different interface/port parameters.

The problem was that symlinks should always be relative, but this is
only appropriate if the original script and its symlink are part of the
same checkout.

The new option allows to achieve a similar effect without symlinking,
searching the original scripts in $PATH instead of hard coding an
absolute (or relative) path to a completely unrelated checkout.