Allowing multiple --override-script-name options
Only the last option is honored.

This allows the user to manually specify that option on the command
line, even if she does not call the script directly but rather via a
wrapper script which also specifies the --override-script-name option.

In such cases the user option overrides the option from the wrapper
script, because the user argments are normally passed last.