Choose appropriate revision info file suffixes
Previously, "_revno.txt" was added to all files. However, this looks
silly then the result is "this is a file_revno.txt".

Therefore, if spaces are detected with in the file name after stripping
the extension, " - revno.txt" will be added instead. This looks then as
follows:

"this is a file - revno.txt"

which is *way* better.

And if you don't like this new heuristic - well, it's always possible to
used -f instead of -F and specify the name of the revision information
file yourself.