cscope wrapper now assumes permanent database files by default
cscope-tmprun-for-tree was written with the idea in mind that most of
the time one wants to browse source files without cluttering the source
directory with database files, and was willing to set up a database
directory for that reason. Or that it is enough to create the database
only for a single run.

Both assumptions have turned out to be wrong in actual usage.

Most of the time, I browse source files temporarily extracted from
archives.

And while they are extracted, I like to run cscope multiple times,
execting other commands like complex grep searches in between.

Therefore, persistency is the default now, and a -t option has been
added in case only a temporary database is wanted.

The script has also been renamed to reflect this changed default
behavior.

The -q option was removed. I do not see a reason for what is could be
good any more, as it was the default anyway.

Finally, online documentation has been added.