Adopt TCP send/receive scripts to new ifconfig
The new ifconfig version outputs its information using a different
format. Parse the output in a way that is compatible with the old as
well as the new format.

In addition, ifconfig seems no longer to be restricted to /sbin.

Therefore perform a normal $PATH search for the executabe, but prepend
$PATH with known superuser directories.