Add passfile- and GNU switch handling support
The password can now be stored as a single-line text file, and the
path to that file can be specified with the --pass-file option rather
than entering it interactively.

Improved switch handling support allows switch clustering with and
without arguments, as well as "--option=arg" as an alternative way
to specify "--option arg".