Add replacement script for apcalc
The script uses yacas to evaluate the expression, which gives pretty
much the same result for simple expressions as using apcalc.

However, yacas is smaller than apcalc, yet much more powerful. It can
certainly be expected to be somewhat slower than apcalc, but this will
only matter when the programming features of apcalc are used (which are
incompatible to those of yacas anyway).