Remove superfluous BusyBox wrapper script
I realized that instead of a wrapper script, one could install a symlink
to (or a copy of) a custom version of BusyBox under the same name just
as easily with less overhead.

I also disliked the underscore in the filename. Therefore, all scripts
which use it, will be updated to use the name
"busybox-pbyqxzl1ktqlk3fjm3arlrclg" and a customized BusyBox executable
with this name will be put in one of the directories in $PATH.