Use different option for inplace backups
The problem was that getopts does not distinguish between

$ comment -i.bak file

and

$ comment -i .bak file

unless it is written as

$ comment -i -- .bak file

Therefore,

$ comment -i file1 file2

did *not* do what one was expecting: file1 was actually used as backup
suffix for file2!

In order to make the difference more clear, and because -i.bak is used
rarely anyway, use two different options instead: -I takes an argument
and -i doesn't. Otherwise, both options do the same thing.