Add script for calculating file sizes from a list
The script reads a list of file system object path names from standard
input and extracts the paths which actually refer to files, ignoring
symlinks, directories and other non-file objects.

It then displays the combined size of all the files from the list.

When output is redirected to a file, only prints the combined size in
bytes.

When standard output is a terminal, it formats the size into an
appropriate unit such as MiB which is easier to read by humans.

This script is especially useful for piping a list of the contents of
some installed software-package to it, because it then calculates the
bytes occupied by the package (except for the file-system overhead which
cannot be determined easily in a generic fashion).