Add scripts for version bumping and C string edits
"bump_version" increments a date-based or serial version number using a
heuristic to determine the flavor used. Can be used either as a filter
or by specifying a list of files to be examined for version bumping. In
addition to version numbers, the copyright end year will also be bumped
if necessary. No changes are made if an input file does not contain a
recognized version number or copyright end year which needs to be
bumped.

"cstr" converts between a multiline C-string consisting of concatenated
line strings, and a plain-text version of it. Again, a heuristic is used
in order to determine the current input format. This is a filter.