Add script for running a command in a terminal
This is intended to be used by graphical shells for executing single
commands in a terminal window without closing the terminal window
automatically, so the user can see the result of the command.

This version is only very rudimentary. It only supports a single
command, i. e. no shell-constructs such as for, while etc.

But hopefully, it handles quoted strings correctly.