Add script for stripping common prefixes/suffixes
It can either detect and remove common pre- and suffixes from the input,
or re-apply them.

This script is useful for editing quoted/indented strings of programming
languages, transforming them into a text-only format which can be
re-flowed as free text and then be transformed back into their original
format.