Update unwrap script and adopt flowtext converter
The unwrap script now considers lines not beginning with something that
automatic line wrapping typically produces to be a separate paragraph,
even without an empty line in between.

The RFC 3676 flowtext converter has also to be modified in order to
comply with this new definition what a paragraph is.

As a consequence the generated markup has to be changed, in order to
make sure it will be treated as a separate paragraph by the unwrap
script.

The new markup starts its meta-instruction lines with closing curly
brace characters instead of opening square brackets. (The latter one
could also be normal text in a paragraph's continuation line.)