Add plaintext formatting to flowtext converter
Unlike -f, the new -t mode does not generate RFC 3676 compliant output,
but rather normal formatted plaintext without trailing whitespace.

Note that this transformation is not reversible. But it is typically
what is wanted when pasting into a MUA which handles RFC 3676 internally
rather than presenting it to the user.