Add another checkpoint utility
This one just writes the checkpoint data to standard output and it quite
flexible in what attributes to include within the checkpoint calculation
what what not to include.

Also, there are not just one hash per checkpoint, but rather two hashes:
One over the selectable attributes, another one over the most basic
attributes like symlink target or file contents, which are always
included depending on the file type.

Both hashes are prefixed by indicators about the data that has been
included within the hash.

This are either the selected optional attributes, of the filesystem
object type.

The fact that contents and attributes are hashed separately allows to
compare file or symlink contents independently from the attributes such
as permissions or owner.