Add script for splitting & checksumming a file
It creates a subdirectory, populating with the contents of the file
split in 650 (by default) pieces.

Then it creates a checksum file over the pieces in the parent directory
of the subdirectory.

Use this for retransmitting damaged pieces of a large file: First, use
the script to chop both source original and destination copy into
pieces. Then compare the checksums and re-transmit damaged parts.
Finally, re-assemble the destination file from the pieces.