Add script to "asciify" locale-encoded text
It is a filter written in Perl. By default, it reads arbitrary text from
standard input or from the files specified as arguments. The text is
assumed to use the current locale's character encoding. Then that text
is converted into pure ASCII, replacing non-ASCII characters by
UNICODE character names enclosed in angle brackets.

This transformation can be undone losslessly be applying the filter
script again, but this time with the option -u.