Check for hex escapes first when decoding UNICODE
It might be safer for asciify_locale-encoded_text to check for
hexadecimal escapes first in case an attacker exploits the fact that
some named UNICODE character has a name that looks like a hex constant.
This is more likely than the other way around, because there are more
possible hex numbers than character names.