Read AWK script via file descriptor
This should be more efficient than passing a large string argument on
the command line, because there maximum length allowed for command line
arguments is restricted and a longer script may exceed the limit.

Redirect via file descriptor # 5 using here-doc syntax. The /dev/fd/5
construct is assumed to provide access to that file descriptor.

I am not sure how portable /dev/fd actually is, but it works on both
Linux and NetBSD which is good enoug for me.