Add PuTTY support and calculate entropy demand
OpenSSH private and public key files will always be generated, but the
new option -p will also create a copy of the private key in PuTTY
format.

Calculate the required amount of entropy to fetch from /dev/random based
on the number of bits required for finding starting points for prime
search, and add a constant hard-coded overhead factor to this in order
to be on the safe side.

Add additional entropy for the last openssl invocation, allowing it to
write out new seeding files with the remaining entropy for the benefit
of future invocations of openssl (i. e. try to be a good citizen, never
leaving behind entropy-exhausted openssl seeding files).

Also make writing unencrypted private key files the default, because I
use it more often. Option -e enables encryption of private key files.