Fix description of middle/inner nested instances
The middle instance normally runs on the local machine, and only the
inner instance runs on the remote machine.

This better isolates remote users than when using the alternative, which
would be to run the middle and inner instances all on the remote
machine, using a single SSH session to that machine.

However, this alternative has the problem that the middle instance
should then be run by user root or at least a fully trusted user.
Because otherwise the unprivileged user running the middle instance
could use it to impersonate any other user which can be switched to
using the middle instance, most likely escalating the unprivileged
user's privileges.