Add option -d for enabling the disk cache
Normally it seems a good idea to bypass the cache, as it is faster and
avoids writes to your SSD, prolonging its life.

However, when enabling the loading of graphics, a disk cache can be
beneficial to reduce network bandwidth usage during multiple sessions,
because the memory cache can obviously not survive a single session.