Remove customization for "visual bell"
This feature turned actually out too annoying, especially when
triggered by tab-completion: The status line was flickering because of
the visual bell all the time.

On the other hand, sometimes it might still be useful.

So it is better to not assume anything and leave the choice up to the
user. Don't enable or disable "visual bell"; just don't touch it and
leave it as it was before.