Don't try to kill the own process group
Although this seems to work when the script is run as a single command,
it kills the other commands in the same pipeline when run as part of a
pipeline.

Instead, wait for the individual child processes to terminate normally
and start a timeout reaper process which kills the child if this takes
too long.

Also support multiple independent pipelines by specifying a pipeline
identifier.

However, currently there is only support for a single pipelines_wait
function which waits for all of the pipelines to finish (or to time
out).