Add command line scanning helper script
Reads the next page from the scanner and converts the received date into
a 300 DPI color scan PNG.

Option -g scans in grayscale instead.

There are also options to override several defaults.

The current page number is written into a file in the current directory.
Delete it to restart at page # 1, or specify the page number directly
with the -N option (which overrides the page number file).