Use -h rather then --help to check for applet
The current method for checking of the existence of a particular BusyBox
applet has been to invoke it with the --help option.

However, BusyBox can also be built without support for long options, and
in this case only -h will work.

Therefore, being the least common denominator, change the applet
checking method to always use -h.