Change BusyBox applet probing algorithm
The method of probing for an applet using -h instead of --help does not
work either, because the applets do not support options -h (or even -H)
at all. At least not as a replacement for --help.

The most generic remaining method of applet dectection is to just invoke
the BusyBox executable and scan its output for the name of the desired
applet.

So do that.

Also change the probing order and not just text the first BusyBox
executable found for the applet, but rather all of them.

Add option -b to specify an alternative BusyBox implementation which
will be probed first.

Include 'busybox-site' as another hard-coded alternative name which will
automatically be probed for.