Migrate to new tmux version of Debian-10
This new version as a preview on by default when choosing a window. This
is nice locally but (I presume) it generates more traffic than
without a preview when run over SSH.

Therefore, add the "-r" option to disable the preview by default.

Also add key bindings for activity and silence monitoring.

Unfortunately I could not find an easy way to toggle silence monitoring,
so the "GNU screen"-compatible shortcut enables it, and a new one turns
of off again.