Scan helper generated .tif files instead of .png
The reason is that PNG compression is much slower than TIF compression
when applied to typical output files of a paper scan. (At least with
ImageMagick, which does not use an optimized PNG encoder.)

Also use lock files now in order to avoid multiple background
invocations of the same script to cause problems. The $counter file
itself is used for locking. Which means the lock is per-directory only.

Finally, add checks to the required external helper utilities and give
advice which packages might need to be installed otherwise.