Don't add a trailing newline at EOF
Partially revert the changes made in the last commit: If the last file
of the input file happens not to be terminated with a newline character,
then don't add one either when writing the transformed line.

This should be most useful when the scripts are invoked with parts of a
line as selections from within an editor which can pipe the selection
through an external filter program.