Add another "joe" configuration for non-wrapping
Even though joe-tab-79 did the right thing in most situations, its
built-in "safety feature" to hard-wrap lines after column 79 which helps
the human detect overlong lines actually made it dangerous to be used
with files containing overlong lines by design.

The new joe-tab-nowrap script is nearly the same as joe-tab-79, but
disables hard word-wrapping, making it suitable for editing the
beforementioned files.

Also add joe-flowed-66 which is the same as joe-flowed-79 except for a
line length of 66 columns. I prefer this line length as most
human-friedly.