Revise escaping algorithm
I am no longer sure whether it is at all possible to properly quote
unwanted characters in a text without parsing that text strictly
left-to-right, but rather performing substitutions with "sed" regular
expressions or "tr".

In order to make sure all transformations can be undone, use a simpler
but less efficient probabilistic algorithm in order to be on the safe
side.

That is, substitute a UUID rather than any clever escape-character
pattern for the unwanted character.

As it is extremely unlikely that the same UUID will appear in the input
text, that method should be very safe.