Add 2 more scripts
"less-human" (symlinked to "more-human" for a more positive-sounding
wording) uses "fold" to fold the input files (or standard input) into
human-friendle 66-column lines and then uses "less" (or "more" if not
installed; both can be overridden by setting $PAGER) to display the
formatted result.

"symlinks-reposess-unowned" searches (using "find") for symbolic links
in the current directory and its subdirectories which have no known user
or group. In those cases, it re-creates the symbolic link as new with
the same target. It will then have the current user as its owner/group
(of course only if the access permissions allow it).