Script to generate script for generating a script
When "script-prefix-prefix" is run with the name of a programming
language as its only argument, it outputs a skeleton script for creating
scripts in this programming language.

The generated script must now be edited in an editor and the actual
shell script prefix for running a script in the intended programming
language must be added at its end.

When the edited script is then chmod +x and run, it will output only the
section which has just been added, replacing any occurrences of the
string $VERSION in the output with the version number found in the
header comments of the script (which by themselves will not be among the
written lines).

Or, explaining it the other way around: If you have a working script
like

> #! /bin/sh
> exec perl -x -- "$0" ${1+"$@"}
> !perl
>
> print "Hello, world!\n"

for running a Perl script, then the lines before the last line are (in
this case) the script prefix.

The last line is the actual Perl script.

Separate now the prefix (the first 4 lines) from the rest and save it
into a file "perl.prefix".

Save the actual script (the last line) as "somescript.pl".

You can now make every Perl script into a shell script by running

$ cat perl.prefix somescript.pl > somescript
$ chmod +x somescript

You might find that very useful and want to do it quite often.
Unfortunately, you need perl.prefix to have around for this. Therefore,
it might be better to have a script "perlscript-prefix" which prints out
the contents of "perl.prefix". The above steps can then be replaced with
this:

$ { perlscript-prefix; cat somescript.pl; } > somescript
$ chmod +x somescript

However, writing "perlscript-prefix" is also a bit of work. The script
"script-prefix-prefix" can be used to generate "perlscript-prefix" like
this:

$ { script-prefix-prefix Perl; cat perl.prefix; } > perlscript-prefix
$ chmod +x perlscript-prefix