Add more comfort to the prefix script generators
Both "cscript-prefix" and "script-prefix-prefix" (which has generated
the "cscript-prefix") now support an optional additional argument which
must be the pathname of an existing file (otherwise the argument will be
ignored).

It this is the case, the file' contents will be output after the prefix,
making the output a complete shell script.

Alternatively, both scripts check whether standard input is a terminal.
If not, both assume a file has been redirected as standard input, and
will output its contents after the prefix they write.

This allows to simplify

$ { script-prefix-prefix Perl; cat perl.prefix; } > perlscript-prefix
$ chmod +x perlscript-prefix

to either

$ script-prefix-prefix Perl perl.prefix > perlscript-prefix
$ chmod +x perlscript-prefix

or

$ script-prefix-prefix Perl < perl.prefix > perlscript-prefix
$ chmod +x perlscript-prefix

Note that the original variant still works, too. It's just an
enhancement.