Avoid cluttering the current directory
"surf" seems to have a large problem: Whenever one surfs to a web site
using it, "surf" downloads a copy of the remote file to the current
directory before viewing it.

This even applies to local files specified with "file://"-URLs!

Create a temporary directory and "cd" there after making any relative
paths specified as arguments into absolute ones.

This will create the copies in the temporary directory.