Add support for word-wrapping and non-break space
The comment in the "script-prefix-prefix" will no longer be copied
through to the output file with only substituting the programming
language name.

Instead, the comments will be re-flown and word-wrapped for 80 column
display. When prefixing a separating space in the comments with a
backslash, word-wrapping will be suppressed at that word boundary.