Add script for decompressed git backup/restore
This script converts a git repository to a single uncompressed flat file
similar to fast-export.

However, it will export not just the current branch, but all local
branches, tags, notes and the default branch setting.

Another difference is that all objects will be exported in binary form,
even tree objects. This is more efficient than the text representation
used by "fast-export".

After creating the export file, it is recommended to compress it using

$ lrzip -L 9 -U -z BACKUP_FILENAME

because this has shown to produce the smallest backups (so far).