Add new radix base 2 through 62 encoder
It encodes into base 62 by default, but a different radix base can be
specified as the only argument.

It is only supposed to be used for rather short inputs such as key
files, because otherwise the calculation effort will increase
at least quadratically.

The new version is implemented in portable AWK.