Upgrade secure SSH key generator
Because commit

https://github.com/torvalds/linux/commit/30c08efec8884fb106b8e57094baa51bb4c44e32

ruined /dev/random as a reliable source of random key material, I
updated my secure version of ssh-keygen by combining three sources of
random numbers for key creation: /dev/urandom, haveged and (if present)
bytes from the hardware random number generator /dev/hwrng.

If only one of those sources generates random numbers with entropy, the
generated keys should be secure.

Actually, haveged alone should alway generate such random numbers with
full entropy, but it might have implementation flaws or overestimate the
amount of entropy it returns. Therefore, mixing with bytes obtained from
/dev/urandom will make it as least as secure as the latter one, which
the original ssh-keygen will also use. Which means, in the worst case we
are no better than the original, but never inferior to that.

One problem however is that the current Debian 11 version of haveged has
a bug so that it cannot be used with the new ssh-keygen-openssl version
if another instance of haveged is running.

If you encounter this problem, you can resolve it as follows:

1. Upgrade to a newer version of haveged which does no longer have that
bug.

2. Temporarily stop the haveged daemon. Then the script will propably
work.

3. Build a private version of haveged and install the patch I have sent
to upstream as well as to the Debian Mailing list. Basically it just
changes a single source line.

Here is a copy:

--- haveged-1.9.14.orig/src/haveged.c
+++ haveged-1.9.14/src/haveged.c
@@ -429,7 +429,7 @@ int main(int argc, char **argv)
       close(socket_fd);
       return ret;
       }
-   else {
+   else if (!(params->setup & RUN_AS_APP)){
       socket_fd = cmd_listen(params);
       if (socket_fd >= 0)
          fprintf(stderr, "%s: command socket is listening at fd %d\n", params->daemon, socket_fd);