Remove newline from text copied to clipboard
Even though a newline is helpful if a new line shall be inserted
somewhere else, it is a problem when copying the pathname of a file to
the clipboard to be pasted into an "open file" dialog, becaus the
newline will be translated into a trailing space which needs to be
manually removed.

Therefore the newline is no longer added by the front end utility - it
is easy to add a newline manually after pasting if required.