Support default argument for random keygen
Also change the order of haveged invocation, avoiding to run it in
parallel if /dev/urandom also triggers reading from it (via a haveged
daemon service).

Note that the Debian 11 Version of haveged has a bug which does not
allow it from being used as daemon and user space program at the same
time. Build it yourself from the upstream sources, or apply a patch to
the Debian version and build it that way.

Only a single line needs to be changed, so it is not a big change.