Provide XML format identification header
It turns out "xmlstarlet" drops the XML header, but retains XML
processing instructions.

Inject a faked processing instruction which will look the same as the
format header for a standalone document.

Provide option "-d" for injecting a different header or leaving it off
completely by specifying no argument to "-d".