Add script for shuffling an *.m3u playlist
Such playlists can be read and written, for instance, by the VLC media
player.

The script pseudo-randomly shuffles the playlist entries, completely
determined by a seed specified by the user. No actual randomness is
involved at all.

This means that the user can regenerate the same resulting output
playlist by specifying the same seed and input playlist.

The user can also create differently shuffled versions of the same
playlist easily be using an incrementing counter as the seed.

The reason why I wrote this script is the fact that the VLC player has
no option of shuffling its playlist. It only has an option to randomly
select the next song from the current playlist. Which is not the same
thing.

I wanted a way to play all the songs in my playlist in seemingly random
order, but a way that allows to quit the player at any time and resume
playback later.

This is now possible with the new script: Firstly, one uses the VLC GUI
in order to construct an initial playlist containing all the songs which
shall be played back in pseudo-random order.

Then this playlist must be saved using the .m3u playlist format.

Next, the new script is used in order to create a shuffled version of
the playlist.

This playlist is then loaded into VLC, and played sequentially from the
start.

Now VLC playback can be quit at any time.

When VLC is run again, the same shuffled playlist can be loaded, and
the user can resume playback where it has been stopped earlier.

In order to resume with the right song, there are 2 strategies:

1. The user can wrote down somewhere at what song to resume next

2. The user can delete all songs already played back from the playlist
and replace the saved playlist with the shortened new version. In this
case, it might be a good idea to keep the original version of the
playist (before shuffling) as well as saving a seed counter used to
derive the last shuffled playlist in a file. This allows the user to
either recreate the current shuffled playlist, as well as increment the
seed counter for generating the next version of a shuffled playlist.