Consolidate "joe" wrapper scripts
All scripts are replaced by symlinks to a single version.

The main reason was the fact that I notices the scripts did not work as
expected if more than a single file has being used as arguments.

In that case, the options injected into the command line by the script
only affected the first file, but not the other ones.

The new consolidated script contains a loop which injects the option
before every file which is passed as an argument.

Also, it used helper functions so that common options do not need to be
specified by every symlink-specific configuration section.