Add a simple script for recreating files
It emits a shell script for recreating a list of files the pathnames of
which will be read from standard input.

This is very much like "shar", but with extreme little overhead. It can
also just handle the most basic cases, that is recreate regular files
and executable script.

No support for symlinks, hardlinks, binary files.