Add new scripts for duplicating/reformatting stdin
"dup" is like cat, but duplicates the output a specified number of times
(default: twice).

"foldwrap" converts between a space-prefixed and a space-suffixed line
encoding of free running paragraph text which has been wrapped for
storage in order to comply with some maximum line length convention.

The space-suffixed format is similar to t he MIME type
"text=plain; format=flowed".

The space-prefixed format is similar to RFC-822 folding of e-mail header
lines, except that the spaces are part of the text and have not been
artificially inserted.

Both formats have their benefits: The prefixed format is preferable in
scripts where trailing spaces are cumbersome.

The suffixed format is directly supported by the reformatting function
of some text editors such as "joe" (when enabling options -flowed
-wordwrap).

The idea is to use an editor's "filter selected block through external
command"-feature to convert the prefixed into the suffixed format first,
then edit the text, and finally convert it back into the prefixed format
which is actually expected by the script.